강화 기능 완성 및 강화한 능력치 세이브

+++ Assets/Scripts/UI/GoldBarUI.cs
@@ -29,7 +29,7 @@ public class GoldBarUI : MonoBehaviour
     /// <summary>
     /// CurrencyManager OnGoldChanged 액션 실행시 실행
     /// </summary>
-    private void UpdateGoldUI(long currentGold)
+    private void UpdateGoldUI(double currentGold)
     {
         _goldText.text = CurrencyFormatter.Format(currentGold);
     }
